+++ NEWS.md
@@ -1,7 +1,7 @@
 # v. 0.1.3
   
   * Simplified default output behaviour
-  * `h3_to_parent()` and `h3_to_children()` added.
+  * `h3_to_parent()`, `h3_to_children()`, `h3_get_kring()`, `h3_get_kring_list()`, and `h3_get_ring()` added.
 
 # v. 0.1.2
 
@@ -11,7 +11,7 @@
 # v. 0.1.1
  
   * `h3_to_geo()` added.
-  * 'geo_to_h3()` bugfix. A pox on devs who think y,x is ok.
+  * `geo_to_h3()` bugfix. A pox on devs who think y,x is ok.
   * NEWS and README added.
 
 # v. 0.1.0